Correct units and buttons
OWM asks for units, Kelvin, Imperial or metric then fails to return rain and show matching the request.  It only returns mm for rain and snow.  I was not making the correction after receipt of the data.  I thought they would do it right.  They are NOT doing it right by ignoring the user request.  But it is documented now so they will never change it.  Sad...  Also button function follows orientation now.